+++ src/core_modules/capture-core/components/Pages/NewRelationship/RegisterTei/GeneralOutput/WarningsSection/SearchGroupDuplicate/ReviewDialogContents.container.js
@@ -7,21 +7,7 @@ import { ReviewDialogContentsComponent } from './ReviewDialogContents.component'
 import withLoadingIndicator from '../../../../../../../HOC/withLoadingIndicator';
 import withErrorMessageHandler from '../../../../../../../HOC/withErrorMessageHandler';
 import { makeDataElementsSelector } from './reviewDialogContents.selectors';
-import type { CardDataElementsInformation, SearchResultItem } from '../../../../../Search/SearchResults/SearchResults.types';
-
-type OwnProps = {|
-    onLink: (id: string, values: any)=>void
-|}
-
-type PropsFromRedux = {|
-    ready: boolean,
-    isUpdating: boolean,
-    error: string,
-    teis: Array<SearchResultItem>,
-    dataElements: CardDataElementsInformation,
-|}
-
-export type Props = {| ...OwnProps, ...PropsFromRedux, ...CssClasses |}
+import type { Props, OwnProps } from './ReviewDialogContents.types';
 
 const makeMapStateToProps = () => {
     const dataElementsSelector = makeDataElementsSelector();
