+++ src/core_modules/capture-core/components/Pages/NewRelationship/RegisterTei/GeneralOutput/WarningsSection/SearchGroupDuplicate/ReviewDialogContentsPager.types.js
@@ -1,0 +1,13 @@
+// @flow
+
+export type OwnProps = {|
+    nextPageButtonDisabled: boolean,
+|};
+type DispatchersFromFromRedux = {|
+    onChangePage: Function,
+|};
+type PropsFromRedux = {|
+    currentPage: number,
+|};
+
+export type Props ={| ...OwnProps, ...DispatchersFromFromRedux, ...PropsFromRedux, ...CssClasses |}
