+++ src/core_modules/capture-core/components/Pages/NewRelationship/RegisterTei/GeneralOutput/WarningsSection/SearchGroupDuplicate/WarningMessageCreator.component.js
@@ -4,7 +4,7 @@ import { withStyles } from '@material-ui/core/styles';
 import i18n from '@dhis2/d2-i18n';
 import LinkButton from '../../../../../../Buttons/LinkButton.component';
 import { ResultsPageSizeContext } from '../../../../../shared-contexts';
-import type { Props } from './WarningMessageCreator.container';
+import type { Props } from './WarningMessageCreator.types';
 
 const getStyles = () => ({
     linkButton: {
