+++ src/core_modules/capture-core/components/Pages/NewRelationship/RegisterTei/GeneralOutput/WarningsSection/SearchGroupDuplicate/WarningMessageCreator.types.js
@@ -1,0 +1,8 @@
+// @flow
+export type OwnProps = {|
+    onOpenReviewDialog: () => void,
+|}
+type DispatchersFromRedux = {|
+    onReviewDuplicates: (onOpenReviewDialog: Function) => void,
+|}
+export type Props = {| ...DispatchersFromRedux, ...OwnProps, ...CssClasses |}
