+++ src/core_modules/capture-core/components/Pages/NewRelationship/RegisterTei/RegisterTei.types.js
@@ -1,0 +1,18 @@
+// @flow
+type PropsFromRedux = {|
+    tetName: ?string,
+    ready: boolean,
+    error: string,
+    possibleDuplicates: ?boolean,
+|};
+type DispatchersFromRedux = {|
+    onReviewDuplicates: (pageSize: number) => void,
+|};
+
+export type OwnProps = {|
+    onLink: (teiId: string) => void,
+    onGetUnsavedAttributeValues?: ?Function,
+    onSave: Function,
+|};
+
+export type Props = {|...PropsFromRedux, ...OwnProps, ...DispatchersFromRedux, ...CssClasses |}
