+++ src/core_modules/capture-core/components/TeiSearch/TeiSearch.component.js
@@ -8,7 +8,7 @@ import { TeiSearchResults } from './TeiSearchResults/TeiSearchResults.container'
 import SearchProgramSelector from './SearchProgramSelector/SearchProgramSelector.container';
 import { Section, SectionHeaderSimple } from '../Section';
 import { ResultsPageSizeContext } from '../Pages/shared-contexts';
-import type { Props } from './TeiSearch.container';
+import type { Props } from './TeiSearch.types';
 
 const getStyles = (theme: Theme) => ({
     container: {
