+++ src/core_modules/capture-core/components/TeiSearch/TeiSearch.container.js
@@ -11,33 +11,7 @@ import {
     setOpenSearchGroupSection,
 } from './actions/teiSearch.actions';
 import { makeSearchGroupsSelector } from './teiSearch.selectors';
-import { type SearchGroup } from '../../metaData';
-
-
-type PropsFromRedux = {|
-    searchGroups: ?Array<SearchGroup>,
-    showResults?: ?boolean,
-    selectedProgramId: ?string,
-    selectedTrackedEntityTypeId: ?string,
-    openSearchGroupSection: ?string,
-|}
-
-type DispatchersFromRedux = {|
-    onSearch: Function,
-    onSearchValidationFailed: Function,
-    onSetOpenSearchGroupSection: (searchId: string, searchGroupId: ?string) => void,
-    onSearchResultsChangePage: (searchId: string, pageNumber: number) => void,
-    onNewSearch: (searchId: string) => void,
-    onEditSearch: (searchId: string) => void,
-|}
-
-type OwnProps = {|
-    id: string,
-    getResultsView: Function,
-    resultsPageSize: number,
-|}
-
-export type Props = {| ...OwnProps, ...DispatchersFromRedux, ...PropsFromRedux, ...CssClasses |}
+import type { Props, OwnProps } from './TeiSearch.types';
 
 const makeMapStateToProps = () => {
     const searchGroupsSelector = makeSearchGroupsSelector();
